Reformat extract class isTaskDue method logic
