+++ src/Command/ScheduledTaskCommand.php
@@ -172,62 +172,12 @@ class ScheduledTaskCommand extends Command
 
     private function validateTask(array $task): array
     {
-        return (new TaskValidator())->validateTask($task);
+        return (new TaskValidator($this->logService))->validateTask($task);
     }
 
     private function isTaskDue(array $task): bool
     {
-        if (
-            $this->isTaskPastStartDate($task)
-            && $this->isTaskNotPastEndDate($task)
-            && $this->isTaskNotExceededMaxExecutions($task)) {
-            $cron = CronExpression::factory($task['expression']);
-
-            return $cron->isDue();
-        }
-
-        return false;
-    }
-
-    private function isTaskPastStartDate(array $task): bool
-    {
-        $start = $task['start'];
-        if (null === $start) {
-            return true;
-        }
-
-        $now = DateHelper::date();
-        $startDate = DateHelper::date($start);
-
-        return $startDate <= $now;
-    }
-
-    private function isTaskNotPastEndDate(array $task): bool
-    {
-        $end = $task['stop'];
-        if (null === $end) {
-            return true;
-        }
-
-        $now = DateHelper::date();
-        $endDate = DateHelper::date($end);
-
-        return $endDate >= $now;
-    }
-
-    private function isTaskNotExceededMaxExecutions(array $task): bool
-    {
-        if (!$this->isLoggingEnabled() || null === $task['times']) {
-            return true;
-        }
-
-        $scheduledTask = $this->logService->getLatestScheduledTaskLog($task['name']);
-
-        if (null === $scheduledTask) {
-            return true;
-        }
-
-        return !$scheduledTask->isRemainingZero();
+        return (new TaskValidator($this->logService))->isTaskDue($task);
     }
 
     private function executeTask(array $task, bool $isAsync, OutputInterface $output): void
